Move apps/translations/ to mkt/translations/ (bug 1021204)
