+++ apps/amo/tests/__init__.py
@@ -35,13 +35,11 @@ from waffle import cache_sample, cache_switch
 from waffle.models import Flag, Sample, Switch
 
 import amo
+import mkt
 from amo.urlresolvers import get_url_prefix, Prefixer, reverse, set_url_prefix
 from constants.applications import DEVICE_TYPES
 from files.models import File, Platform
 from lib.post_request_task import task as post_request_task
-from translations.models import Translation
-
-import mkt
 from mkt.access.acl import check_ownership
 from mkt.access.models import Group, GroupUser
 from mkt.constants import regions
@@ -50,10 +48,11 @@ from mkt.feed.indexers import (FeedAppIndexer, FeedBrandIndexer,
 from mkt.files.helpers import copyfileobj
 from mkt.prices.models import AddonPremium, Price, PriceCurrency
 from mkt.site.fixtures import fixture
+from mkt.translations.models import Translation
 from mkt.users.models import UserProfile
 from mkt.versions.models import Version
-from mkt.webapps.models import update_search_index as app_update_search_index
 from mkt.webapps.models import Addon, Category, Webapp, WebappIndexer
+from mkt.webapps.models import update_search_index as app_update_search_index
 from mkt.webapps.tasks import unindex_webapps
