+++ mkt/submit/tests/test_views.py
@@ -17,18 +17,18 @@ import amo.tests
 import mkt
 from amo.tests import formset, initial
 from amo.tests.test_helpers import get_image_path
-from mkt.users.models import UserNotification
-from mkt.users.notifications import app_surveys
 from constants.applications import DEVICE_TYPES
 from mkt.files.tests.test_models import UploadTest as BaseUploadTest
 from mkt.site.fixtures import fixture
 from mkt.submit.decorators import read_dev_agreement_required
 from mkt.submit.forms import AppFeaturesForm, NewWebappVersionForm
 from mkt.submit.models import AppSubmissionChecklist
+from mkt.translations.models import Translation
+from mkt.users.models import UserNotification
+from mkt.users.models import UserProfile
+from mkt.users.notifications import app_surveys
 from mkt.webapps.models import (Addon, AddonCategory, AddonDeviceType,
                                 AddonUser, AppFeatures, Category, Webapp)
-from translations.models import Translation
-from mkt.users.models import UserProfile
 
 
 class TestSubmit(amo.tests.TestCase):
