+++ mkt/translations/forms.py
@@ -1,10 +1,10 @@
 from django.conf import settings
 from django.db import models
 from django.forms.util import ErrorList
-from django.utils.translation.trans_real import to_language
 from django.utils.encoding import force_unicode
-from django.utils.safestring import mark_safe
 from django.utils.html import conditional_escape
+from django.utils.safestring import mark_safe
+from django.utils.translation.trans_real import to_language
 
 
 def default_locale(obj):
