+++ mkt/translations/models.py
@@ -8,6 +8,7 @@ import commonware.log
 import amo
 import amo.models
 from amo import urlresolvers
+
 from . import utils
