+++ mkt/translations/templates/translations/all-locales.html
