+++ mkt/translations/transformer.py
@@ -2,8 +2,8 @@ from django.conf import settings
 from django.db import connections, models, router
 from django.utils import translation
 
-from translations.models import Translation
-from translations.fields import TranslatedField
+from mkt.translations.fields import TranslatedField
+from mkt.translations.models import Translation
 
 isnull = """IF(!ISNULL({t1}.localized_string), {t1}.{col}, {t2}.{col})
             AS {name}_{col}"""
